requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning & Computer Vision
# PyTorch nightly with CUDA 12.8 support for RTX 5070 (sm_120)
# Install with: pip install --pre torch torchvision torchaudio --index-url https://download.pytorch.org/whl/nightly/cu128
torch
torchvision
torchaudio
# Install ultralytics after PyTorch to avoid version conflicts
ultralytics
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
scipy==1.11.4

# HuggingFace & Transformers
transformers==4.36.0
huggingface-hub==0.19.4
tokenizers==0.15.0
accelerate==0.25.0
safetensors==0.4.1

# Data Processing
pandas==2.1.3

# Configuration & Environment
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# WebRTC & Streaming
aiortc==1.6.0
python-socketio==5.10.0
websockets==12.0

# Async & Concurrency
aiofiles==23.2.1
httpx==0.25.2
aiohttp==3.9.1

# Utilities
tqdm==4.66.1
typing-extensions==4.8.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Optional: Database support
# psycopg2-binary==2.9.9
# sqlalchemy==2.0.23
# alembic==1.12.1